Move memory to impl
